requirements.txt
gradio>=4.0
openai>=1.0
pillow
python-dotenv
